// types/product.ts
export interface ProductType {
  name: string;
  category: string;
  rent_category: number;
  rnb: boolean;
  weight: number;
  height: number;
  width: number;
  length: number;
  daily_price: number;
  weekly_price: number;
  monthly_price: number;
  description: string;
  condition_description: string;
  stock: number;
  status: string;
  image: string;
  address: string;
  rating: number;
  rented_times: number;
}
  